add test for page with no day, time, and location data
This test for `Extractor` contains a page that has, for its class details, a single row for a class with no information for the Days and Location columns, and a single dash for the Time column.

With no additional processing, the current version of the program fails on this test, with the output of the `Extractor` converting the empty-content cells into `na`.