Change expected N/A type
To accomodate for possibly empty values within other columns of the class details table (and not just Dates, Days, Location), "No schedule" has been changed to "none". This maintains a custom N/A data value, so that any possible unexpected N/A data values produced by pd.read_html() can be easily distinguished and detected.

Along with this, additional expected values have been added to the regex for the Size and Available columns.