Add report draft
This report visualises statistics from the compiled data for the current day. Being an RMarkdown report, this can be done for any day by passing a value to the `date` parameter, taking the format "YYYY-MM-DD".

It may be rendered by running `rmarkdown::render` whilst within the `results/` directory of this project.