Change na_rep to 'none' for compiled course details
To ensure that all 'na' types are consistent throughout the compiled class and course details, the 'na' entries obtained throughout compilation of the course details table are changed to 'none' upon conversion to CSV.

This may potentially obscure errors within the data, however (e.g. when 'none' appears in an unexpected place in the data).