Add pre-processing for course details tables
There is one page (CRARTS 3003) which currently contains an incorrect entry in the "Biennial Course" column of the course details table (this entry should instead be in the "Assessment" column). Thus appropriate pre-processing has been added in this commit.

This also makes tests.compiled_data pass (at least for the current dataset).