Add detected of unexpected NA data
As is evident from the generation of the report, some of the pages still contain asubstantial amount of missing information, which is being interpreted as NA upon extraction with pandas. This NA data has not been explicitly changed as `none` anywhere in the extraction process, so it is unexpected NA data.

This commit adds two ways of detecting unexpected NA data: one through Extractor.__sanitise_data() and another through tests.compile_data.