Add check for warnings
For each page in the current dataset, all warnings currently appear together at a particular location of the page. The added code in this commit checks for warnings.

Redundant regex matching for 'Warning' found within the code for gathering annotations has also been removed, since it fails to detect warnings.