Fix skipping of scrapes in Scraper
If the user entered 'n' in the yes/no prompt for `scrape_url()` to skip the scrape, the program will not crash but the invalid scraped page gets saved. Also, if the user entered 'n' in the prompt within `get_course_list()` to skip the scrape, the program crashes since the page scraped was invalid. This commit patches a fix for that.