Add validation for course listing page
After obtaining the links to the course lists for each subject, the Scraper visits scrapes each of those pages. These pages currently have no validation, and due to the current workings of the code, the Scraper will skip scraping course lists it cannot scrape from. This commit contains additions for page validation to fix this, prompting the user to repeat the scrape upon an invalid page.

The Scraper also had no warning output upon an invalid page, so code to output an appropriate warning is included in this commit to ensure that an indication is given for when the course list page is invalid.