Quote subject area when scraping course list page
One subject, namely `C&ENVENG`, has been ignored for most of the scrapes. This is because the scraper substitutes the subject area name into a template URL to create the course list link, but since this subject contains an ampersand character, the incorrect parameter is passed. This commit fixes this by ensuring the percent-encoding of the subject area name before substituting it into the template URL.