Modify verification for empty course lists
Since the verification for a course list page only checks whether any links to courses can be found, the scraper considers subjects with no courses as invalid. In the current dataset, one subject area contains an empty course list, so this commit has an adjustment to the verification to also ensure that at least the <table> containing potential course links is present. It does this by checking whether at least one <th class="course"> is present (if such a tag is present, then the course list should be present).