Implement compile_to_sql()
This commit contains the addition of an implementation for compile_to_sql() in src/prepare.py. This implementation uses the pandas.to_sql and pandas.read_sql methods as much as possible, inserting the merge of the course details and class details tables to a SQLite database under data/store.db.